+++ addons/confirma/src/attributes/AfterAllAttribute.cs
@@ -2,5 +2,9 @@ using System;
 
 namespace Confirma.Attributes;
 
-[AttributeUsage(AttributeTargets.Method)]
-public class AfterAllAttribute : Attribute { }
+[AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
+public class AfterAllAttribute : LifecycleAttribute
+{
+    public AfterAllAttribute(string methodName = "AfterAll")
+    : base(methodName) { }
+}