+++ addons/confirma/src/attributes/BeforeAllAttribute.cs
@@ -2,5 +2,9 @@ using System;
 
 namespace Confirma.Attributes;
 
-[AttributeUsage(AttributeTargets.Method)]
-public class BeforeAllAttribute : Attribute { }
+[AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
+public class BeforeAllAttribute : LifecycleAttribute
+{
+    public BeforeAllAttribute(string methodName = "BeforeAll")
+    : base(methodName) { }
+}