+++ addons/confirma/src/attributes/LifecycleAttribute.cs
@@ -1,0 +1,18 @@
+using System;
+
+namespace Confirma.Attributes;
+
+[AttributeUsage(
+    AttributeTargets.Class,
+    Inherited = true,
+    AllowMultiple = false
+)]
+public class LifecycleAttribute : Attribute
+{
+    public string MethodName { get; init; }
+
+    public LifecycleAttribute(string className)
+    {
+        MethodName = className;
+    }
+}