+++ CHANGELOG.md
@@ -44,6 +44,7 @@ All notable changes to this project will be documented in this file.
 - New settings window
 - [#178] Allow changing category via editor
 - Allow disabling parallelization via editor
+- Base class for lifecycle attributes (LifecycleAttribute).
 
 ### Changed
 
@@ -56,12 +57,15 @@ All notable changes to this project will be documented in this file.
 - GdTestDiscovery & GdTestExecutor now use GdScriptInfo instead of ScriptInfo.
 - GDScript tests folder can be passed as global or localized path.
 - `Print` and `PrintLine` methods from Log class support list of arguments.
-- Rewritten docs for arguments.
 - Move verbose switch to new settings window
+- From now on, lifecycle attributes are assigned to a class,
+not to a specific method, and take the name of the method to run.
+One attribute of a given type is allowed.
 
 ### Removed
 
 - GdAssertionFailed signal.
+- LifecycleMethodData record.
 
 ### Fixed